fix building time bugs and errors

+++ src/services/product/server.ts
@@ -1,13 +1,15 @@
-import { PRODUCT_DIR } from '@/src/constants/url';
+import { productFilePath } from '@/src/utils/filePathProvider';
 import fs from 'fs'
-import path from 'path';
 
-export const deleteProductImageService = async (filePath:string) => {
+export const deleteProductImageService = async (filePath: string) => {
   try {
     if (!filePath) return
-    const fileFullPath = path.join(PRODUCT_DIR, filePath)
+    const fileFullPath = productFilePath(filePath)
+    if (!fileFullPath) {
+      return
+    }
     if (!fs.existsSync(fileFullPath)) return
-     await fs.rmSync(fileFullPath)
+    await fs.rmSync(fileFullPath)
   } catch (err) {
     console.error(err);
   }
